More fixes

+++ test/regression/aarch64/Exception.cc
@@ -292,7 +292,7 @@ TEST_P(Exception, svcr) {
     smstop
     smstart
   )");
-  for (int i = 0; i < (SVL / 8); i++) {
+  for (uint64_t i = 0; i < (SVL / 8); i++) {
     CHECK_MAT_ROW(ARM64_REG_ZA, i, uint32_t, fillNeon<uint32_t>({0}, SVL / 8));
   }
